pass fsyscalltest part 3

+++ kern/include/syscall.h
@@ -64,7 +64,7 @@ int sys_close(int fd);
 int sys_read(int fd, void *buf, size_t buflen, int *retval);
 int sys_write(int fd, void *buf, size_t nbytes, int *retval);
 
-int sys_lseek(int fd, off_t pos, int whence, int *retval);
+int sys_lseek(int fd, off_t pos, int whence, int64_t *retval);
 
 int sys_chdir(const char *pathname);
 int sys_dup2(int oldfd, int newfd, int *retval);
